+++ src/common/init.c
@@ -315,6 +315,12 @@ static void defaultConfig(FFinstance* instance)
     instance->config.gpuTemp = false;
     instance->config.batteryTemp = false;
 
+    instance->config.gpuHideIntegrated = false;
+    instance->config.gpuHideDiscrete = false;
+
+    instance->config.shellVersion = true;
+    instance->config.terminalVersion = true;
+
     instance->config.titleFQDN = false;
 
     ffStrbufInitA(&instance->config.diskFolders, 0);