+++ CHANGELOG.md
@@ -15,7 +15,11 @@ Features:
 * Add `Brightness` module support
 * Add `Battery` module support for FreeBSD
 * Add `--disk-show-unknown` option for Disk module
+* Add `--gpu-hide-integrated` option (#379)
+* Add `--gpu-hide-discrete` option (#379)
 * Detect terminal version when available
+* Support `WezTerm` terminal font detection (requires [`wezterm` executable](https://wezfurlong.org/wezterm/cli/general.html) being available)
+* Add `--shell-version` and `--terminal-version` options to disable shell / terminal version detection
 
 Logos:
 * Raspbian (@IamNoRobot, #373)
@@ -23,6 +27,7 @@ Logos:
 Bugfixes:
 * `--logo-type` now does accept `iterm` too (@CarterLi, #374)
 * Fix mintty terminal font detection
+* Fix bug that line buffering doesn't work properly (Windows)
 
 Other:
 * Fixed a Typo in iterm error message (@jessebot, #376)